Create secrets backends shared module
