+++ airflow-core/src/airflow/config_templates/config.yml
@@ -1594,8 +1594,8 @@ secrets:
         .. note:: |experimental|
 
         Comma-separated list of secret backends. These backends will be used in the order they are specified.
-        Please note that the `environment_variable` and `metastore` are required values and cannot be removed
-        from the list. Supported values are:
+        Please note that the ``environment_variable`` and ``metastore`` are required values and cannot be
+        removed from the list. Supported values are:
 
         * ``custom``: Custom secret backend specified in the ``secrets[backend]`` configuration option.
         * ``environment_variable``: Standard environment variable backend
